XML request does not validate against the Notifier 2.3 XSD schema
Link to schema: http://airbrake.io/airbrake_2_3.xsd

The generated XML includes a 'component'  attribute in each <line> node
in the backtrace. Per the schema, this is not a valid attribute and the
generated XML failed to validate. At the time of testing, said responses
would yield a 200, rather than the expected 422, response from the
Notifier API, however would never show up in Airbrake. The URL included
with the response returns "Please retry. This notice UUID is still being
processed." indefinitely. Airbrake was notified of this issue and
included the generated XML.

To correct the issue, AirBrake.makeBackTraceLine() no longer allows for
an optional component argument and thus does not include that attribute
in the <line> node. To continue to send this value to the <component>
node of the <request> node, that value is parsed from the stack trace
separately.