XML request doesn't always include required 'url' node on 'request' node
If null is sent as a URL to AirBrakeXML.makeNotice(), no 'url' node
will be added to the 'request' node. Per the Notifier API 2.3 specs,
this is a required node in a 'request' (though the 'request' node
itself is optional).

This commit creates an empty 'url' node if the url argument is null.