Record errors from non-debug Flash Players
The AirBrake API requires a <component> node to be included in the
<request> node. Because non-debug players do not include a stack
trace, the attempt to determine the component will return null
resulting in that node not being sent to Airbrake and the error
not being recorded.

This change defaults the component to 'unknown' thereby allowing
the resulting XML to be valid.